Started to move rewards-related helpers into their own module
