+++ contracts/mixnet/src/rewards/mod.rs
@@ -1,2 +1,3 @@
+pub mod helpers;
 pub mod queries;
 pub mod transactions;
